将"稀疏码的at"做为"概念的dt",以修复内类比中"不同算法区,产生有无" (如:"吃的有-无") 的BUG;

+++ SMG_NothingIsAll/AIFoundation/AIHeader.h
@@ -138,4 +138,4 @@ va_end(args);\
  */
 #define DefaultAlgsType @" "    //默认AlgsType
 #define DefaultDataSource @" "  //默认DataSource
-#define AlgNodeAlgsType(pId) STRFORMAT(@"%d",pId)   //概念节点AlgsType
+#define AlgNodeAlgsType(pId) STRFORMAT(@"%ld",(long)pId)   //概念节点AlgsType
